+++ src/main/java/com/ost/matie/repository/product/ProductRepositoryCustom.java
@@ -5,6 +5,6 @@ import com.ost.matie.domain.product.Product;
 import java.util.List;
 
 public interface ProductRepositoryCustom {
-    List<Product> findAllByCategoryId(Integer categoryId);
+    List<Product> findAllByCategoryId(Long categoryId);
     List<Product> findAllByNameContaining(String name);
 }
